// edelphi/src/main/typescript/src/generated/client/models/PanelUserGroup.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eDelphi REST API
 * REST API for eDelphi
 *
 * The version of the OpenAPI document: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PanelUserGroup
 */
export interface PanelUserGroup {
    /**
     * Id
     * @type {number}
     * @memberof PanelUserGroup
     */
    readonly id?: number;
    /**
     * 
     * @type {string}
     * @memberof PanelUserGroup
     */
    name: string;
}

export function PanelUserGroupFromJSON(json: any): PanelUserGroup {
    return PanelUserGroupFromJSONTyped(json, false);
}

export function PanelUserGroupFromJSONTyped(json: any, ignoreDiscriminator: boolean): PanelUserGroup {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': json['name'],
    };
}

export function PanelUserGroupToJSON(value?: PanelUserGroup | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
    };
}